(#46) Folder restructure
