+++ src/client/index.scss
@@ -2,7 +2,7 @@
 @import 'styles/vars';
 @import 'styles/base';
 
-@import 'view/components/Header';
+@import 'view/Header';
 
 @import 'view/index';
