+++ src/client/view/Editor/EditorHeader.js
