+++ src/client/view/Editor/getHeaderButtons.js
