+++ src/client/view/Editor/index.js
