+++ src/client/view/Icons.js
