+++ src/client/view/Sidebar/SidebarHeader.js
