+++ src/client/view/Sidebar/index.js
