+++ src/client/view/Tags/index.js
